requirements.txt
# Core tools
typer[all]
rich

# Formatting, linting, testing
black
ruff
pytest
pytest-cov

# Optional: For .env support
python-dotenv

# Configuration
pyyaml
